refactor(ObserverLocator): remove general purpose ElementObserver
Initially ElementObserver was used only when EventManager returned a handler (inputs, selects, textareas).  Subsequent releases expanded the ElementObserver to include support for other scenarios such as data-* attributes, xlink: attributes, style, etc.  When this was done the ObserverLocator was changed to use the ElementObserver for *all* Element properties, which caused issue #57.  This refactor breaks the ElementObserver into single-responsibility observers for each of the Element scenarios we handle (style/css, select-value, input-checked, data-*, aria-*, xlink:, etc).  All other scenarios fall through to the standard property observation logic (OoPropertyObserver, SetterObserver, etc).

Fixes #57
