+++ src/observer-locator.js
@@ -6,11 +6,16 @@ import {DirtyChecker, DirtyCheckProperty} from './dirty-checking';
 import {
   SetterObserver,
   OoObjectObserver,
-  OoPropertyObserver,
-  ElementObserver,
+  OoPropertyObserver
+} from './property-observation';
+import {
   SelectValueObserver,
   CheckedObserver,
-} from './property-observation';
+  ValueAttributeObserver,
+  XLinkAttributeObserver,
+  DataAttributeObserver,
+  StyleObserver
+} from './element-observation';
 import {All} from 'aurelia-dependency-injection';
 import {
   hasDeclaredDependencies,
@@ -128,7 +133,7 @@ export class ObserverLocator {
   }
 
   createPropertyObserver(obj, propertyName){
-    var observerLookup, descriptor, handler, observationAdapter;
+    var observerLookup, descriptor, handler, observationAdapter, xlinkResult;
 
     if(obj instanceof Element){
       handler = this.eventManager.getElementHandler(obj, propertyName);
@@ -138,7 +143,19 @@ export class ObserverLocator {
       if (propertyName ==='checked' && obj.tagName.toLowerCase() === 'input') {
         return new CheckedObserver(obj, handler, this);
       }
-      return new ElementObserver(obj, propertyName, handler);
+      if (handler) {
+        return new ValueAttributeObserver(obj, propertyName, handler);
+      }
+      xlinkResult = /^xlink:(.+)$/.exec(propertyName);
+      if (xlinkResult) {
+        return new XLinkAttributeObserver(obj, propertyName, xlinkResult[1]);
+      }
+      if (/^\w+:|^data-|^aria-/.test(propertyName) || obj instanceof SVGElement) {
+        return new DataAttributeObserver(obj, propertyName);
+      }
+      if (propertyName === 'style' || propertyName === 'css') {
+        return new StyleObserver(obj, propertyName);
+      }
     }
 
     descriptor = Object.getPropertyDescriptor(obj, propertyName);
